Move MatchScorer in with the others
